Testing completed payment on stripe side

+++ frontend/client/src/app/checkout/checkout-payment/checkout-payment.component.ts
@@ -23,6 +23,7 @@ export class CheckoutPaymentComponent implements OnInit {
   cardNumber?: StripeCardNumberElement;
   cardExpiry?: StripeCardExpiryElement;
   cardCvc?: StripeCardCvcElement;
+  cardErrors: any;
 
   constructor(private basketService: BasketService, private router: Router,
     private checkoutSvc: CheckoutService, private toastr: ToastrService) { }
@@ -34,10 +35,30 @@ export class CheckoutPaymentComponent implements OnInit {
       if(elements) {
         this.cardNumber = elements.create('cardNumber');
         this.cardNumber.mount(this.cardNumberElement?.nativeElement);
+        this.cardNumber.on("change", event => {
+          if(event.error) this.cardErrors = event.error.message;
+          else {
+            this.cardErrors = null;
+          }
+        })
+
         this.cardExpiry = elements.create('cardExpiry');
         this.cardExpiry.mount(this.cardExpiryElement?.nativeElement);
+        this.cardExpiry.on("change", event => {
+          if(event.error) this.cardErrors = event.error.message;
+          else {
+            this.cardErrors = null;
+          }
+        })
+
         this.cardCvc = elements.create('cardCvc');
         this.cardCvc.mount(this.cardCvcElement?.nativeElement);
+        this.cardCvc.on("change", event => {
+          if(event.error) this.cardErrors = event.error.message;
+          else {
+            this.cardErrors = null;
+          }
+        })
       }
     })
   }
@@ -51,9 +72,22 @@ export class CheckoutPaymentComponent implements OnInit {
     this.checkoutSvc.createOrder(orderToCreate).subscribe({
       next: createdOrder => {
         this.toastr.success("Order created successfully");
-        this.basketService.deleteLocalBasket();
-        const navigationExtras: NavigationExtras = {state: createdOrder}
-        this.router.navigate(['checkout/success'], navigationExtras);
+        this.stripe?.confirmCardPayment(basket.clientSecret!, {
+          payment_method: {
+            card: this.cardNumber!,
+            billing_details: {
+              name: this.checkoutForm?.get('paymentForm')?.get('nameOnCard')?.value
+            }
+          }
+        }).then(result => {
+          if(result.paymentIntent) {
+            this.basketService.deleteLocalBasket();
+            const navigationExtras: NavigationExtras = {state: createdOrder}
+            this.router.navigate(['checkout/success'], navigationExtras);
+          } else {
+            this.toastr.error(result.error.message)
+          }
+        });
       }
     })
   }
